Extrais le gestionnaire d'actions tirroir
